+++ hari_client/models/models.py
@@ -984,6 +984,37 @@ class MLAnnotationModel(BaseModel):
     )
 
 
+class MLAnnotationModelMetrics(pydantic.BaseModel):
+    ml_model_id: uuid.UUID = pydantic.Field(
+        title="Id of ML Model the metrics belong to"
+    )
+    category: str = pydantic.Field(
+        title="Category of the metrics (either overall for all data or per answer option)"
+    )
+    test_curves: dict[str, list[float]] = pydantic.Field(
+        title="Automation Degree Correctness data for automation, correctness and confidence for test set"
+    )
+    val_curves: dict[str, list[float]] | None = pydantic.Field(
+        default=None,
+        title="Automation Degree Correctness data for automation, correctness and confidence for validation set",
+    )
+    precomputed_curves: dict[str, list[float]] | None = pydantic.Field(
+        default=None,
+        title="Precomputed automation, correctness and confidence for cut off threshold calculation. "
+        "Includes validation accuracy and confidence thresholds, test accuracy and automation degree.",
+    )
+    created_at: datetime.datetime | None = pydantic.Field(
+        default=None, title="Created At"
+    )
+    archived_at: datetime.datetime | None = pydantic.Field(
+        default=None, title="Archived At"
+    )
+
+
+class MLAnnotationModelWithMetrics(MLAnnotationModel):
+    metrics: list[MLAnnotationModelMetrics] | None = None
+
+
 class AIAnnotationRun(BaseModel):
     id: uuid.UUID = pydantic.Field(title="Id")
     created_at: datetime.datetime | None = pydantic.Field(